+++ src/meltano/core/setting_definition.py
@@ -129,8 +129,13 @@ class SettingDefinition(NameEq, Canonical):
     def is_redacted(self):
         return self.kind in ("password", "oauth")
 
-    def env_vars(self, namespace: str):
-        env_keys = [self.env or to_env_var(namespace, self.name)]
+    def env_vars(self, namespaces: [str]):
+        env_keys = []
+
+        if self.env:
+            env_keys.append(self.env)
+
+        env_keys.extend(to_env_var(namespace, self.name) for namespace in namespaces)
 
         env_keys.extend(alias for alias in self.env_aliases)