+++ tests/meltano/core/test_elt_context.py
@@ -32,7 +32,9 @@ class TestELTContext:
         with invoker.prepared(session):
             env = invoker.env()
 
-        assert env["TAP_MOCK_TEST"] == "mock"
+        # Extractor
+        assert env["MELTANO_EXTRACT_TEST"] == env["TAP_MOCK_TEST"] == "mock"
+        assert env["MELTANO_EXTRACT__SELECT"] == env["TAP_MOCK__SELECT"] == '["*.*"]'
 
     def test_loader(self, elt_context, session, target_postgres):
         loader = elt_context.loader
@@ -43,13 +45,24 @@ class TestELTContext:
         with invoker.prepared(session):
             env = invoker.env()
 
+        # Extractor
         assert env["MELTANO_EXTRACTOR_NAME"] == elt_context.extractor.name
         assert env["MELTANO_EXTRACTOR_NAMESPACE"] == elt_context.extractor.namespace
-        assert env["MELTANO_EXTRACT_TEST"] == "mock"
-        assert env["TAP_MOCK_TEST"] == "mock"
 
-        assert env["PG_ADDRESS"] == os.getenv("PG_ADDRESS", "localhost")
-        assert env["PG_SCHEMA"] == env["MELTANO_EXTRACTOR_NAMESPACE"]
+        assert env["MELTANO_EXTRACT_TEST"] == env["TAP_MOCK_TEST"] == "mock"
+        assert env["MELTANO_EXTRACT__SELECT"] == env["TAP_MOCK__SELECT"] == '["*.*"]'
+
+        # Loader
+        assert (
+            env["MELTANO_LOAD_HOST"]
+            == env["PG_ADDRESS"]
+            == os.getenv("PG_ADDRESS", "localhost")
+        )
+        assert (
+            env["MELTANO_LOAD_SCHEMA"]
+            == env["PG_SCHEMA"]
+            == env["MELTANO_EXTRACTOR_NAMESPACE"]
+        )
 
     def test_transformer(self, elt_context, session, dbt):
         transformer = elt_context.transformer
@@ -60,21 +73,45 @@ class TestELTContext:
         with invoker.prepared(session):
             env = invoker.env()
 
+        # Extractor
         assert env["MELTANO_EXTRACTOR_NAME"] == elt_context.extractor.name
         assert env["MELTANO_EXTRACTOR_NAMESPACE"] == elt_context.extractor.namespace
-        assert env["MELTANO_EXTRACT_TEST"] == "mock"
-        assert env["TAP_MOCK_TEST"] == "mock"
 
+        assert env["MELTANO_EXTRACT_TEST"] == env["TAP_MOCK_TEST"] == "mock"
+        assert env["MELTANO_EXTRACT__SELECT"] == env["TAP_MOCK__SELECT"] == '["*.*"]'
+
+        # Loader
         assert env["MELTANO_LOADER_NAME"] == elt_context.loader.name
         assert env["MELTANO_LOADER_NAMESPACE"] == elt_context.loader.namespace
-        assert env["MELTANO_LOAD_HOST"] == os.getenv("PG_ADDRESS", "localhost")
-        assert env["MELTANO_LOAD_SCHEMA"] == env["MELTANO_EXTRACTOR_NAMESPACE"]
-        assert env["PG_ADDRESS"] == os.getenv("PG_ADDRESS", "localhost")
-        assert env["PG_SCHEMA"] == env["MELTANO_EXTRACTOR_NAMESPACE"]
 
-        assert env["DBT_TARGET"] == env["MELTANO_LOADER_NAMESPACE"]
-        assert env["DBT_TARGET_SCHEMA"] == "analytics"
-        assert env["DBT_SOURCE_SCHEMA"] == env["MELTANO_EXTRACTOR_NAMESPACE"]
+        assert (
+            env["MELTANO_LOAD_HOST"]
+            == env["PG_ADDRESS"]
+            == os.getenv("PG_ADDRESS", "localhost")
+        )
+        assert (
+            env["MELTANO_LOAD_SCHEMA"]
+            == env["PG_SCHEMA"]
+            == env["MELTANO_EXTRACTOR_NAMESPACE"]
+        )
+
+        # Transformer
+        assert (
+            env["MELTANO_TRANSFORM_TARGET"]
+            == env["DBT_TARGET"]
+            == env["MELTANO_LOADER_NAMESPACE"]
+        )
+        assert (
+            env["MELTANO_TRANSFORM_TARGET_SCHEMA"]
+            == env["DBT_TARGET_SCHEMA"]
+            == "analytics"
+        )
+        assert (
+            env["MELTANO_TRANSFORM_SOURCE_SCHEMA"]
+            == env["DBT_SOURCE_SCHEMA"]
+            == env["MELTANO_EXTRACTOR_NAMESPACE"]
+        )
+        assert env["MELTANO_TRANSFORM_MODELS"] == env["DBT_MODELS"]
         assert env["DBT_MODELS"].startswith(env["MELTANO_EXTRACTOR_NAMESPACE"])
 
     def test_select_filter(self, elt_context, session):